requirements.txt
# v2 매매 프로그램 의존성
PyYAML>=6.0
openpyxl>=3.0.0
pytz>=2023.3
MetaTrader5>=5.0.45
python-dotenv>=1.0.0
requests>=2.28.0
psutil>=5.9.0
Pillow>=9.0.0
matplotlib>=3.5.0
pyautogui>=0.9.50
# ktr_lots 웹 스크래핑 사용 시에만 필요
# selenium>=4.0.0
